+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 
 - The commands `push` and `pull` to transmit snapshots
+- The `keep`command to remove snapshots
 
 ### Changed